LPS-55800 Move TemplateRenderer tag
