+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-taglib/src/META-INF/resources/taglib/ui/template_renderer/init-ext.jspf
@@ -13,5 +13,3 @@
  * details.
  */
 --%>
-
-<%@ page import="com.liferay.portlet.dynamicdatamapping.DDMTemplate" %>
