+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-taglib/src/com/liferay/dynamic/data/mapping/taglib/ui/TemplateRendererTag.java
@@ -12,15 +12,15 @@
  * details.
  */
 
-package com.liferay.taglib.ddm;
+package com.liferay.dynamic.data.mapping.taglib.ui;
 
+import com.liferay.dynamic.data.mapping.taglib.ui.base.BaseTemplateRendererTag;
 import com.liferay.portal.kernel.portletdisplaytemplate.PortletDisplayTemplateManagerUtil;
 import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.theme.ThemeDisplay;
 import com.liferay.portal.util.PortalUtil;
 import com.liferay.portlet.dynamicdatamapping.DDMTemplate;
 import com.liferay.taglib.FileAvailabilityUtil;
-import com.liferay.taglib.ddm.base.BaseTemplateRendererTag;
 
 import java.util.HashMap;
 import java.util.Map;
